+++ src/server/helpers/proxy-request.ts
@@ -1,4 +1,4 @@
-import * as request from 'request';
+import request from 'request';
 import { fromNode } from './build-response';
 import { KrasInjectorInfo } from '../types';